Sprint 3 Code
This is all of the Sprint 3 code that I have written.  I checked and it seems like everything works just fine.  There are some txt files that aren't included, but you should just be able to create them easily.  It is imperative that you integrate these files into your own project, as I made a lot of changes in order to implement my solutions.